Cleanup and consolidation of JS target
Moved compiler and runner (cmd) into dedicated directory; eventually
this will be the pattern for all targets. Additionally, I extracted out
the shared binding functions (externs and builtins) into libraries which
can be imported into a "harness" script. This greatly streamlines the
test runner and also sets up nicely for running on node + web.
